third_party/WebKit/LayoutTests/fast/dom/dom-attribute-on-prototype-chain.html

<!DOCTYPE>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
var div = document.createElement("div");
shouldBeFalse('div.hasOwnProperty("id")');
shouldBeTrue('div.__proto__.__proto__.__proto__.hasOwnProperty("id")');
shouldThrow('xhr = new XMLHttpRequest(); xhr.__proto__ = HTMLDivElement.prototype; xhr.id');
shouldThrow('obj = Object.create(div); obj.id');
</script>
</body>
</html>
